Move some functions into and out of utils.js
