+++ source/features/show-recently-pushed-branches-on-more-pages.tsx
@@ -9,7 +9,7 @@ This feature also adds this widget to the Issues List, Issue page and PR page
 import {React} from 'dom-chef/react';
 import select from 'select-dom';
 import features from '../libs/features';
-import {getRepoURL} from '../libs/page-detect';
+import {getRepoURL} from '../libs/utils';
 
 const repoUrl = getRepoURL();
