+++ source/features/show-user-top-repositories.tsx
@@ -6,7 +6,7 @@ https://user-images.githubusercontent.com/1402241/48474026-43e3ae80-e82c-11e8-93
 import {React} from 'dom-chef/react';
 import select from 'select-dom';
 import features from '../libs/features';
-import {getCleanPathname} from '../libs/page-detect';
+import {getCleanPathname} from '../libs/utils';
 
 function init() {
 	const showcaseTitle = select('.js-pinned-repos-reorder-container .text-normal');
