+++ test/page-detect.js
@@ -14,74 +14,6 @@ function urlMatcherMacro(t, detectFn, shouldMatch = [], shouldNotMatch = []) {
 	}
 }
 
-test('getRepoPath', t => {
-	const pairs = new Map([
-		[
-			'https://github.com',
-			false
-		],
-		[
-			'https://gist.github.com/',
-			false
-		],
-		[
-			'https://github.com/settings/developers',
-			false
-		],
-		[
-			'https://github.com/sindresorhus/notifications/notifications',
-			false
-		],
-		[
-			'https://github.com/sindresorhus/refined-github',
-			''
-		],
-		[
-			'https://github.com/sindresorhus/refined-github/',
-			''
-		],
-		[
-			'https://github.com/sindresorhus/refined-github/blame/master/package.json',
-			'blame/master/package.json'
-		],
-		[
-			'https://github.com/sindresorhus/refined-github/commit/57bf4',
-			'commit/57bf4'
-		],
-		[
-			'https://github.com/sindresorhus/refined-github/compare/test-branch?quick_pull=0',
-			'compare/test-branch'
-		],
-		[
-			'https://github.com/sindresorhus/refined-github/tree/master/distribution',
-			'tree/master/distribution'
-		]
-	]);
-
-	for (const [url, result] of pairs) {
-		location.href = url;
-		t.is(result, pageDetect.getRepoPath(url));
-	}
-});
-
-test('getOwnerAndRepo', t => {
-	const ownerAndRepo = {
-		'https://github.com/sindresorhus/refined-github/pull/148': {
-			ownerName: 'sindresorhus',
-			repoName: 'refined-github'
-		},
-		'https://github.com/DrewML/GifHub/blob/master/.gitignore': {
-			ownerName: 'DrewML',
-			repoName: 'GifHub'
-		}
-	};
-
-	Object.keys(ownerAndRepo).forEach(url => {
-		location.href = url;
-		t.deepEqual(ownerAndRepo[url], pageDetect.getOwnerAndRepo());
-	});
-});
-
 test('is404', t => {
 	document.title = 'Page not found · GitHub';
 	t.true(pageDetect.is404());
